requirements.txt
azure-identity
azure-mgmt-network
azure-mgmt-resource
graphviz